+++ lib/laflem/collections/macos/homebrew/__init__.py
@@ -1,0 +1,6 @@
+'''
+Import all modules in this package.
+'''
+from .setup import HomebrewSetupModule
+from .upgrade import HomebrewUpgradeModule
+from .cleanup import HomebrewCleanupModule