+++ lib/laflem/exceptions/http.py
@@ -1,0 +1,10 @@
+'''
+Defined HTTP exceptions.
+'''
+
+from .base import FlemException
+
+class HTTPException(FlemException):
+  '''
+  The base HTTP exception.
+  '''